code/requirements.txt
beautifulsoup4
lxml
requests
pandas
rdflib
owlready2
biopython
nltk
lxml